requirements.txt
streamlit==1.32.0
pandas==2.2.1
lightweight-charts==1.0.0
google-generativeai==0.3.2
python-dotenv==1.0.1
streamlit-lightweight-charts==1.0.6 